bi-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
grpcio==1.67.1
protobuf==5.29.5
python-dotenv==1.0.1
